+++ src/StateSmith/Output/UserConfig/RenderConfigAllVars.cs
@@ -21,11 +21,17 @@ public class RenderConfigAllVars
     /// </summary>
     public readonly RenderConfigJavaScriptVars JavaScript;
 
+    /// <summary>
+    /// NOTE! Field name used with reflection for toml parsing.
+    /// </summary>
+    public readonly RenderConfigJavaVars Java;
+
     public RenderConfigAllVars()
     {
         Base = new RenderConfigBaseVars();
         C = new RenderConfigCVars();
         CSharp = new RenderConfigCSharpVars();
         JavaScript = new RenderConfigJavaScriptVars();
+        Java = new RenderConfigJavaVars();
     }
 }