+++ src/test-misc/java-spec2/Spec2SmBase.java
@@ -1,0 +1,9 @@
+// import MainClass;
+
+public class Spec2SmBase
+{
+    protected static boolean trace_guard(String message, boolean b)
+    {
+        return MainClass.traceGuard(message, b);
+    }
+}